Fix booting race conditions
By starting later (at filesystem) we're able to ensure everything we
need is in place. Prior to that there seem to be a subtle relationship
between plymouth and its upstart-bridge which can cause signals to fail.

Additionally we've moved to a new baseami which is from master in the
baseami product.