Use edge keyname generator for nontracked
Summary:
In gologging there are keyname generators, these take a filename and generate an s3 keyname.

It's a bit dubious that they are in gologging but it's out of scope to change that. When Dhruv changed them processor one to respect versions
it broke nontracked files. It was a minor break, the files still ended up in s3 but with a weird keyname.

There is an "edge" keyname generator that is already used by spade for audit log files, so I am just using it also for nontracked files

Test Plan:
- Integration test passed
- Manual review of schema suggestor log files show the names are correct

Reviewers: mperrone

Reviewed By: mperrone

Subscribers: rhu, maasted

Differential Revision: https://phabricator.internal.justin.tv/D3219