processor: fix geoip updater and refactor config
Summary:
https://twitchtv.atlassian.net/browse/SCIENG-265

The goeip updater was broken. There was a bug where the updater gets closed
immediately after starting, which hid the other error. The deeper problem was
that the config that points to the locations in s3 of the dbs was malformed.

I closed the updater at a reasonable time. Factored out a lot of config into the
config object, and added logging so it's clear when the update fails or
succeeds.

Test Plan: integration test

Reviewers: rhu

Reviewed By: rhu

Subscribers: ojcm, dwe, gareth, jackgao, maasted

Differential Revision: https://phabricator.internal.justin.tv/D3500