Adding a health check listener
Summary: Listening and serving port 8080 and replying a status code of 200 if the agent is alive. This makes the agent support ELB style health checks, which are a best practice when using autoscaling groups.

Test Plan: Ran integration test with new AMI and verified that the ELB was able to detect a healthy instance as the service was running, then killed the service and the instance went out, then resumed and the instance came back online

Reviewers: mperrone

Reviewed By: mperrone

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D3553