Create Spark replay scripts and fix final bugs for replay
Summary:
Adds two scripts to the top level, replay.py and replay.sh.  The Python
script is a Spark job which lives on the master Spark node, runs the processor
in replay mode over a selection of Edge records, then filters the results and
reingests them into Ace (after deleting records from Ace from the same
interval).  replay.sh encapsulates this for ease of use, since ssh'ing into
the master node and submitting the job to Spark are both necessary.

Since replays are not part of our normal pipeline and do not trigger the
standard ingester or audit logger, there is no reason for it to use SNS, so I
turned all such messaging off in replay mode.

Spade was not designed to run with multiple instances on the same machine; the
output and log files from separate instances clobber each other.
Unfortunately, there's no convenient way to guarantee in Spark that any worker
node will contain exactly one instance of Spade.  To get around this, we now
build a Docker image as part of building the AMI, and Spark runs this image
instead of running Spade directly.

Finally, we now close the audit logger later in the termination sequence
(since closing the audit logger before other processes are finished logging
can trigger a panic), and make sure that the GZip writers have finished
writing and uploading all the records before closing them.

Test Plan:
Passes a new version of the integration test (which I am about to submit for
review) which replays the data from stage 1 and verifies that the counts in
Ace remain the same.

Reviewers: jackgao, dwe

Reviewed By: jackgao, dwe

Subscribers: dwe, maasted, #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4378