Adding a login to id transformer (part 1)
Summary:
In preparation for name changes being made into a public feature, this commit brings the redshift type userIDWithMapping. To mitigate risk, this brings the minimal set of changes required to just test that the new scoop_protocol and transformer are supported, but no effort to map will be taken in the transformer.

For performance analysis, RedshiftTransformer now expects a StatsLogger to record transforming times.

Additionally, the following are links to required changes in other projects:

- [[ https://phabricator.internal.justin.tv/D4695 | scoop_protocol ]]

Test Plan: Run a modified integration test that sends event that follow the new protocol and use the new transformer and verify it runs fine. On production, deploy and verify grafana stats are not affected in any significant way.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4702