processor: make geoip response string comparison match new 304 format
Summary:
The geoip db in the processor is updated regularly. There is a job on the jobs
server
(https://git-aws.internal.justin.tv/stats/jobs/blob/master/jobs/run_geoip_update.sh) that pulls it from its
source and uploads it to s3 twice a week. The processor polls the s3 location
with a GetObject request and an If-Modified-Since header that makes it respond
304 if it hasn't been updated since it was last downlaoded to the processor.
I just updated aws-sdk-go, and they changed the string format of the 304
response, such that the if statement didn't identify it as a 304. I looked
through the docs again, and it doesn't look like there's a better way to
discover it's a 304 than this string:

- http://docs.aws.amazon.com/sdk-for-go/api/aws/awserr/#Error
- http://docs.aws.amazon.com/sdk-for-go/api/service/s3/#GetObjectOutput

I updated the string, but this could happen again if we update again and they
change the response code. Only thing slightly more robust than this that I
considered was doing:

```
if strings.Contains(err.Code(), "Not") && strings.Contains(err.Code(), "Modified") {
```

Do you think this is better? Is there another way?

Test Plan:
https://git-aws.internal.justin.tv/mperrone/scratch/blob/master/aws-sdk-go/main.go <- this gave me what I expected, hurrah.

I think in order to test this without requiring network talking to the aws api, we'd have to mock the response,
which is the thing that would matters here and would change, so I don't think this can be a unit test.

Going to brainstorm with #secret-scieng on this, so pushing the diff without an added test to reference.

Reviewers: clgroft

Reviewed By: clgroft

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4819