Adding a login to id transformer (part 2)
Summary:
This is the second part to 8abfe78. It brings the rest of the changes required to implement the userIDWithMapping transformer. This transformer sets the value of a column expected to be a user id when is empty/invalid, and for that it expects a supporting column with a login so we can use it to map to an id through an underlying API.

The changes include:

- We now make explicit two types of transformers: single value transformer and mapping transformer. Single value includes all the transformers we had until this point, which just transforms the value provided without help of other columns. Mapping transformer are expected to use supporting columns to map values to the target column. Currently, only the new userIDWithMapping type is a mapping transformer.
- Mapping transformers are expected to be provided a config that tells them about the ValueFetcher and StringCache to use. ValueFetcher are used to fetch values from somewhere outside the transformer’s control and StringCache is used to cache string results to avoid a fetch if possible.
- To fetch a user id, a JSONValueFetcher is provided to the userID transformer. This just makes a GET to get a json from an endpoint and extracts a particular field from it using a JQ expression.
- To cache a user id, an elastimemcache Client is provided. Is basically a wrapper over gomemcache that is able to auto discover nodes with a cluster id.
- To facilitate all this, the following config fields were added and corresponding places updated:
  * JSONValueFetchers: map of id to JSONValueFetcher configurations
  * TransformerCacheCluster: params for interaction with transformer cache
  * TransformerFetchers: map of transformer id to value fetcher id

Test Plan: Run a modified integration with a cache cluster deployed and verify it runs well and transfomer latency is not significantly affected. On prod, manually deploy to a small subset of instances and monitor performance, adjust memcache config is required, and continue rolling out to mor instances.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4795