Use gzrecover to recover partial data from failed gzip writes.
Summary:
when the processor gets terminated without going through the shutdown sequence,
it leaves invalid gzip files, but much of the data can be recovered using the
gzrecover tool. This adds gzrt to the build, and on startup, salvages the files
found in the events folder using the tool, and uploads them.
Also added a unit test. wahoo!

Test Plan:
the unit test passes. I also tested manually on integration thusly:
1 - tmux on the processor, with 3 panes open:
    + `watch -n 1 ls -al /opt/science/spade/data/events`
    + `sudo tail -f /var/log/upstart/spade.log`
    + typed out but not submitted: `sudo kill -9 $(ps -aux | grep spade/bin/spade | grep -v grep | awk '{print $2}'); sudo service spade stop`
2 - start `./integration.py` and wait for events to be sent
3 - see that the `ls` shows files in it with more than just the 10byte gz/file
  header, so files have been written
4 - kill it and keep it dead with the typed out commands
5 - cp the partial mw file to inspect it later
6 - `sudo service spade start`
7 - watch the logs for when it says it uploaded a file to s3 and submitted a file to sqs and copy that s3 URI
8 - dl the .gz file uploaded, gunzip it, and inspect it
9 - use `gzrecover` on the saved mw file, and delete the last line
10 - diff the manually `gzrecover`ed against the one uploaded, and see that they
   are the same!

Before I added the uploader.ClearEventsFolder to main and made it only salvage
then, this test showed that there was junk at the end of the uploaded file, presumably
because a worker was gz appending to the file while `gzrecover` was working.

Reviewers: clgroft

Reviewed By: clgroft

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4954