Add flags to run Spark and DB upload separately
Summary:
This lets us have dry runs where we run the processor but don't upload to the
database (so we don't have to put Blueprint into maintenance mode), as well as
spot-check the Spark output before uploading to Redshift.

Test Plan: Integration test passes.  About to perform a dry run on older data.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4987