Fix args keys in replay and errors in uploader
Summary:
Command-line flags which are written with double dashes in docopt also take
those dashes in the keys to the args hash.

Also, while dry-running replays I was having some trouble diagnosing the
origin of some errors in the uploader; so I added labels to help with that
in the future.

Test Plan:
Integration test succeeds.  In prod I'll redeploy and do another
dry run of the first phase of replays.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4990