Filter events by name in replay driver
Summary:
It's easy for multi-day replays to get prohibitively expensive.  To
mitigate this in the cases when we want events for very few tables,
use a simple filter on the lines we get from the Edge bucket.  Also avoids
getting the filesystem involved, which may also be an efficiency gain.

Test Plan:
New replay script works in integration.  In production, will
run a processor-only replay on invariant_ping and retroactive_complete,
load only invariant_ping into Ace, and make sure the results are what we
expect (and that retroactive_complete isn't just empty or something) before
loading retroactive_complete into Ace.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5019