Remove stringToIntegerMD5
Summary:
stringToIntegerMD5 was originally intended to hash user agents for privacy, and
possibly strings to use as distkeys. It is unused and the ingester doesn't
know it exists, so it's safer to destroy it.

As a result, was also able to delete the hash transformer.

Test Plan: Unit and integration tests pass.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5422