Simplify parsing
Summary:
Remove the fanoutParser and use JSON parser directly.
The fanoutParser was used to migrate from nginx logs to json logs and is no longer necessary.
It also uses global state to track the available parsers, which we should not do if/when we
re-implement this logic.

Test Plan: Ran unit tests and integration test.

Reviewers: clgroft

Reviewed By: clgroft

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5601