Log Kinesis failure counts
Summary:
In the case where PutRecords() or PutRecordsBatch() returns without
an explicit error but fails to put some records on the stream or
Firehose, we were not previously logging the types of failures, which
made debugging difficult.

Test Plan:
Integration test succeeds.  Will look at errors in prod when they
happen again.

Reviewers: mperrone

Reviewed By: mperrone

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5722